attempted to refactor out todoForm to be child of modal, and removed concurrently
